requirements.txt
scikit_image==0.16.2
scipy==1.4.1
torchvision==0.5.0
torch==1.4.0
numpy==1.18.1
Pillow==7.1.2
PyYAML==5.3.1